+++ sentry-ruby/spec/spec_helper.rb
@@ -1,5 +1,6 @@
 require "bundler/setup"
 require "pry"
+require "timecop"
 require 'simplecov'
 require 'rspec/retry'